Use `stash_active_state`
Move `stash_active_state` to misc.utils, and use it more places. Makes
the internal code look neater, and reuses code.
